Added some survey detail prefab

+++ Assets/2.Scripts/2.Model/DTOs/SNSurveyResponseDTO.cs
@@ -8,7 +8,7 @@ public class SNSurveyResponseDTO
     public string Title { get; set; }
     public int TotalQuestion { get; set; }
     public string Status { get; set; }
-    public int Point { get; set; }
+    public float Point { get; set; }
     public string StartDate { get; set; }
     public string ExpiredDate { get; set; }
     public int CreatedUserId { get; set; }
